Starting working on the alien (enemy) first will try to display a single enemy middle of the screen.
Each alien should show on a random point on screen, should have space
between other aliens so they don't overlap. Each alien should be able to
move towards the player (spaceship). When picking a random point, should
ensure that point does not fall over the middle line of the screen just
to give more space for the player to defend.