Creating a new way to display UI elements like start screen and pause mmenu ..etc.
There are so many ways to implement this as its my first time, i was
struggling how to implment it in a way that will ensure easier
reusablity. But I still feel there is 100% another easier way. But for
now will keep it as is. Single UI as an entity and all UI elements are
displayed there.